BinaryTraits

BinaryTraits

Macros

Functions

Types

Macros

BinaryTraits.@trait — Macro

@trait <name> [as <category>] [prefix <positive>,<negative>] [with <trait1,trait2,...>]

Create a new trait type for name called $(name)Trait.

If the as clause is provided, then category (an abstract type) will be

used as the super type of the trait type.

If the prefix clause is provided, then it allows the user to choose different

prefixes than the default ones (Can and Cannot) e.g. prefix Is,Not or prefix Has,Not.

If the with clause is provided, then it defines a composite trait from existing

traits. Note that you must specify at least 2 traits to make a composite trait.

BinaryTraits.@assign — Macro

@assign <T> with <Trait1, Trait2, ...>

Assign traits to the data type T. For example:

    @assign Duck with Fly,Swim

is translated to something like:

    flytrait(::Duck) = CanFly()
    swimtrait(::Duck) = CanSwim()

where x is the name of the trait X in all lowercase, and T is the type being assigned with the trait X.

BinaryTraits.@implement — Macro

@implement <CanType> by <FunctionSignature>

Register function signature for the specified CanType of a trait. You can use the @check macro to verify your implementation after these interface contracts are registered.

For example:

    @implement CanFly by fly(direction::Direction, speed::Float64)

BinaryTraits.@check — Macro

@check <T>

Check whether the data type T fully implements all of its assigned traits. Return an InterfaceReview object.

Functions

BinaryTraits.istrait — Function

istrait(x)

Return true if x is a trait type e.g. FlyTrait is a trait type when it is defined by a statement like @trait Fly.

Types

BinaryTraits.InterfaceReview — Type

InterfaceReview

An InterfaceReview object contains the validation results of an interface.

Missing docstring.

Missing docstring for Assignable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Contract. Check Documenter's build log for details.