BinaryTraits.@assign — Macro

@assign <T> with <Trait1, Trait2, ...>

Assign traits to the data type T. Translated to something like:

    <x>trait(::T) = Can<X>()

where x is the name of the trait X in all lowercase, and T is the type being assigned with the trait X.

BinaryTraits.@check — Macro

@check <Type>

Check whether the data type T fully implements all of its assigned traits. Return a named tuple with the following attributes:

fully_implemented: Bool

missing_contracts: Contract[]

BinaryTraits.@implement — Macro

@implement <CanType> by <FunctionSignature>

Register function signature for the specified CanType of a trait.

BinaryTraits.@trait — Macro

@trait <name> [as <category>] [prefix <positive>,<negative>] [with <trait1,trait2,...>]

Create a new trait type for name called $(name)Trait.

If the as clause is provided, then category (an abstract type) will be

used as the super type of the trait type.

If the prefix clause is provided, then it allows the user to choose different

prefixes than the default ones (Can and Cannot) e.g. prefix Is,Not or prefix Has,Not.

If the with clause is provided, then it defines a composite trait from existing

traits. Note that you must specify at least 2 traits to make a composite trait.

BinaryTraits.istrait — Method

istrait(x)

Return true if x is a trait. This function is expected to be extended by users for their trait types. The extension is automatic when the @trait macro is used.

BinaryTraits.InterfaceReview — Type

InterfaceReview

An InterfaceReview object contains the validation results of an interface.

BinaryTraits.Assignable — Constant

The Assignable type represents any data type that can be associated with traits. For example, a Duck type is Assignable and so it may be assigned with traits CanFly and CanSwim.

The reason why it includes UnionAll is to support parametric types that are not fully qualified e.g. AbstractArray.

BinaryTraits.interface_map — Constant

interface_map

The interface_map is a two-layer Dict data structure. It maps a module to another Dict that maps a data type to a set of Contracts.

For example, a Duck

BinaryTraits.prefix_map — Constant

The prefix_map is a Dict that maps a module to another Dict, which maps a trait (as a Symbol) to tuple of positive and negative prefixes for the trait type. For example:

Main -> :Fly -> (:Can, :Cannot)

BinaryTraits.traits_map — Constant

The traits_map is a two-layer Dict. First layer is to map from a module to data types that have been assigned with traits. The second layer maps a data type to the Can-type of the asigned traits.

For example, a user module ZooKeeper may define Duck and Dog types and then assign them with Fly and Swim traits. The map would look like this:

julia> BinaryTraits.traits_map[ZooKeeper]
Dict{Union{DataType, UnionAll},Set{DataType}} with 2 entries:
  Dog  => Set(DataType[CanSwim])
  Duck => Set(DataType[CanSwim, CanFly])

BinaryTraits.assign — Method

assign(m::Module, T::Assignable, can_type::DataType)

Assign data type T with the specified Can-type of a trait.

BinaryTraits.contracts — Method

contracts(m::Module, can_type::DataType)

Returns a set of Contracts that are required to be implemented for can_type.

BinaryTraits.display_expanded_code — Method

display_expanded_code(expr::Expr)

Display the expanded code from a macro for debugging purpose. Only works when the verbose flag is set using set_verbose.

BinaryTraits.fully_implemented — Method

fully_implemented(m::Module, T::Assignable)

Check if the data type T defined in module m has fully implemented all trait functions that it was previously assigned. See also: @assign.

BinaryTraits.get_prefix — Method

Get trait type positive/negative prefixes for trait.

BinaryTraits.is_tuple_of_symbols — Method

Check if x is an expression of a tuple of symbols. If n is specified then also check whether the tuple has n elements. The op argument is used to customize the check against n. Use >= or <= to check min/max constraints.

BinaryTraits.parse_trait_args — Method

Parse arguments for the @trait macro.

BinaryTraits.register — Method

register(m::Module, can_type::DataType, func::Function,
         args::NTuple{N,DataType}, ret::DataType)

Register a function func with the specified can_type type. The func is expected to take arguments args and return a value of type ret.

BinaryTraits.set_prefix — Method

Set trait type positive/negative prefixes for trait.

BinaryTraits.trait_func_name — Method

trait_func_name(t)

Return the name of the trait instropectin function given a trait t. For example, it would be flytrait for a Fly trait.

BinaryTraits.trait_type_name — Method

trait_type_name(t)

Return the name of trait type given a trait t. For example, it would be FlyTrait for a Fly trait. The trait is expected to be in TitleCase, but this function automatically convert the first character to upper case regardless.

BinaryTraits.traits — Method

traits(m::Module, T::Assignable)

Returns a set of Can-types that the data type T exhibits. See also @assign.

BinaryTraits.Contract — Type

Contract

A contract refers to a function defintion func that is required to satisfy the Can-type of a trait. The function func must accepts args and returns ret.