Reference

Macros

BinaryTraits.@trait — Macro

@trait <name> [as <category>] [prefix <positive>,<negative>] [with <trait1,trait2,...>]

Create a new trait type for name called $(name)Trait:

If the as clause is provided, then category (an abstract type) will be used as the super type of the trait type.

If the prefix clause is provided, then it allows the user to choose different prefixes than the default ones (Can and Cannot) e.g. prefix Is,Not or prefix Has,Not.

If the with clause is provided, then it defines a composite trait from existing traits. Note that you must specify at least 2 traits to make a composite trait.

BinaryTraits.@assign — Macro

@assign <T> with <Trait1, Trait2, ...>

Assign traits to the data type T. For example:

@assign Duck with Fly,Swim

is translated to something like:

flytrait(::Duck) = CanFly()
swimtrait(::Duck) = CanSwim()

where x is the name of the trait X in all lowercase, and T is the type being assigned with the trait X.

BinaryTraits.@implement — Macro

@implement <CanType> by <FunctionSignature>

Register function signature for the specified CanType of a trait. You can use the check macro to verify your implementation after these interface contracts are registered. The function signature only needs to specify required arguments other than the object itself. Also, return type is optional and in that case it will be ignored by the interface checker.

For examples:

@implement CanFly by fly(direction::Direction, speed::Float64)
@implement CanFly by has_wings()::Bool

The data types that exhibit those CanFly traits must implement the function signature with the addition of an object as first argument i.e.

fly(duck::Duck, direction::Direction, speed::Float64)
has_wings(duck::Duck)::Bool

Functions

BinaryTraits.traits — Function

traits(T::Assignable)

Returns a set of Can-types that the data type T exhibits. See also @assign.

BinaryTraits.check — Function

check(T::Assignable)

Check if the data type T has fully implemented all trait functions that it was previously assigned. See also: @assign.

BinaryTraits.required_contracts — Function

required_contracts(T::Assignable)

Return a set of contracts that is required to be implemented for the provided type T.

BinaryTraits.istrait — Function

istrait(x)

Return true if x is a trait type e.g. FlyTrait is a trait type when it is defined by a statement like @trait Fly.

Types

BinaryTraits.InterfaceReview — Type

InterfaceReview

An InterfaceReview object contains the validation results of an interface.

Fields

type: the type being checked

result: true if the type fully implements all required contracts

implemented: an array of implemented contracts

misses: an array of unimplemented contracts

BinaryTraits.Assignable — Constant

Assignable

Assignable represents any data type that can be associated with traits. It essentially covers all data types including parametric types e.g. AbstractArray

BinaryTraits.Contract — Type

Contract{T <: DataType, F <: Function, N}

A contract refers to a function defintion func that is required to satisfy the Can-type of a trait. The function func must accepts args and returns ret.

Fields

can_type: can-type of a trait e.g. CanFly

func: function that must be implemented to satisfy this trait

args: arguments of the function func

ret: return type of the function func